+++ Tests/SentryTests/ViewCapture/SentryViewPhotographerTests.swift
@@ -7,7 +7,7 @@ import XCTest
 
 class SentryViewPhotographerTests: XCTestCase {
     
-    private class TestViewRenderer: ViewRenderer {
+    private class TestViewRenderer: SentryViewRenderer {
         func render(view: UIView) -> UIImage {
             UIGraphicsImageRenderer(size: view.bounds.size).image { context in
                 view.layer.render(in: context.cgContext)